--- code/lib/cli-storybook/src/automigrate/fixes/csf-factories.ts
@@ -1,430 +1,0 @@
-/* eslint-disable no-underscore-dangle */
-import { types as t } from 'storybook/internal/babel';
-import { formatFileContent } from 'storybook/internal/common';
-import {
-  isValidPreviewPath,
-  loadConfig,
-  loadCsf,
-  printConfig,
-  printCsf,
-} from 'storybook/internal/csf-tools';
-
-import picocolors from 'picocolors';
-import prompts from 'prompts';
-
-import type { FileInfo } from '../codemod';
-import { runCodemod } from '../codemod';
-import { detectRenderer } from '../helpers/detectRenderer';
-import {
-  getFrameworkPackageName,
-  getRendererPackageNameFromFramework,
-} from '../helpers/mainConfigFile';
-import type { CommandFix } from '../types';
-
-export async function configToCsfFactory(
-  info: FileInfo,
-  { configType, frameworkPackage }: { configType?: 'main' | 'preview'; frameworkPackage: string },
-  { dryRun = false, skipFormatting = false }: { dryRun?: boolean; skipFormatting?: boolean } = {}
-) {
-  const config = loadConfig(info.source);
-  try {
-    config.parse();
-  } catch (err) {
-    logger.log(`Error when parsing ${info.path}, skipping:\n${err}`);
-    return info.source;
-  }
-
-  configType = configType || (info.path.includes('preview') ? 'preview' : 'main');
-  const methodName = configType === 'main' ? 'defineMain' : 'definePreview';
-
-  const programNode = config._ast.program;
-  const hasNamedExports = Object.keys(config._exportDecls).length > 0;
-
-  /**
-   * Scenario 1: Mixed exports
-   *
-   * ```
-   * export const tags = [];
-   * export default {
-   *   parameters: {},
-   * };
-   * ```
-   *
-   * Transform into: `export default defineMain({ tags: [], parameters: {} })`
-   */
-  if (config._exportsObject && hasNamedExports) {
-    const exportDecls = config._exportDecls;
-
-    for (const [name, decl] of Object.entries(exportDecls)) {
-      if (decl.init) {
-        config._exportsObject.properties.push(t.objectProperty(t.identifier(name), decl.init));
-      }
-    }
-
-    programNode.body = programNode.body.filter((node) => {
-      if (t.isExportNamedDeclaration(node) && node.declaration) {
-        if (t.isVariableDeclaration(node.declaration)) {
-          node.declaration.declarations = node.declaration.declarations.filter(
-            (decl) => t.isIdentifier(decl.id) && !exportDecls[decl.id.name]
-          );
-          return node.declaration.declarations.length > 0;
-        }
-      }
-      return true;
-    });
-  } else if (config._exportsObject) {
-    /**
-     * Scenario 2: Default exports
-     *
-     * - Syntax 1: `default export const config = {}; export default config;`
-     * - Syntax 2: `export default {};`
-     *
-     * Transform into: `export default defineMain({})`
-     */
-    const defineConfigCall = t.callExpression(t.identifier(methodName), [config._exportsObject]);
-
-    let exportDefaultNode = null as any as t.ExportDefaultDeclaration;
-    let declarationNodeIndex = -1;
-
-    programNode.body.forEach((node) => {
-      // Detect Syntax 1
-      if (t.isExportDefaultDeclaration(node) && t.isIdentifier(node.declaration)) {
-        const declarationName = node.declaration.name;
-
-        declarationNodeIndex = programNode.body.findIndex(
-          (n) =>
-            t.isVariableDeclaration(n) &&
-            n.declarations.some(
-              (d) =>
-                t.isIdentifier(d.id) &&
-                d.id.name === declarationName &&
-                t.isObjectExpression(d.init)
-            )
-        );
-
-        if (declarationNodeIndex !== -1) {
-          exportDefaultNode = node;
-          // remove the original declaration as it will become a default export
-          const declarationNode = programNode.body[declarationNodeIndex];
-          if (t.isVariableDeclaration(declarationNode)) {
-            const id = declarationNode.declarations[0].id;
-            const variableName = t.isIdentifier(id) && id.name;
-
-            if (variableName) {
-              programNode.body.splice(declarationNodeIndex, 1);
-            }
-          }
-        }
-      } else if (t.isExportDefaultDeclaration(node) && t.isObjectExpression(node.declaration)) {
-        // Detect Syntax 2
-        exportDefaultNode = node;
-      }
-    });
-
-    if (exportDefaultNode !== null) {
-      exportDefaultNode.declaration = defineConfigCall;
-    }
-  } else if (hasNamedExports) {
-    /**
-     * Scenario 3: Named exports export const foo = {}; export bar = '';
-     *
-     * Transform into: export default defineMain({ foo: {}, bar: '' });
-     */
-    const exportDecls = config._exportDecls;
-    const defineConfigProps = [];
-
-    // Collect properties from named exports
-    for (const [name, decl] of Object.entries(exportDecls)) {
-      if (decl.init) {
-        defineConfigProps.push(t.objectProperty(t.identifier(name), decl.init));
-      }
-    }
-
-    // Construct the `define` call
-    const defineConfigCall = t.callExpression(t.identifier(methodName), [
-      t.objectExpression(defineConfigProps),
-    ]);
-
-    // Remove all related named exports
-    programNode.body = programNode.body.filter((node) => {
-      if (t.isExportNamedDeclaration(node) && node.declaration) {
-        if (t.isVariableDeclaration(node.declaration)) {
-          node.declaration.declarations = node.declaration.declarations.filter(
-            (decl) => t.isIdentifier(decl.id) && !exportDecls[decl.id.name]
-          );
-          return node.declaration.declarations.length > 0;
-        }
-      }
-      return true;
-    });
-
-    // Add the new export default declaration
-    programNode.body.push(t.exportDefaultDeclaration(defineConfigCall));
-  }
-
-  const configImport = t.importDeclaration(
-    [t.importSpecifier(t.identifier(methodName), t.identifier(methodName))],
-    t.stringLiteral(frameworkPackage + `/${configType === 'main' ? 'node' : 'browser'}`)
-  );
-  // only add the import if it doesn't exist yet
-  if (
-    !programNode.body.some(
-      (node) => t.isImportDeclaration(node) && node.source.value === configImport.source.value
-    )
-  ) {
-    programNode.body.unshift(configImport);
-  }
-
-  const output = printConfig(config).code;
-
-  if (dryRun) {
-    logger.log(`Would write to ${picocolors.yellow(info.path)}:\n${picocolors.green(output)}`);
-    return info.source;
-  }
-
-  return skipFormatting ? output : formatFileContent(info.path, output);
-}
-
-export async function storyToCsfFactory(info: FileInfo) {
-  const csf = loadCsf(info.source, { makeTitle: () => 'FIXME' });
-  try {
-    csf.parse();
-  } catch (err) {
-    logger.log(`Error when parsing ${info.path}, skipping:\n${err}`);
-    return info.source;
-  }
-
-  const metaVariableName = 'meta';
-
-  /**
-   * Add the preview import if it doesn't exist yet:
-   *
-   * `import { config } from '#.storybook/preview'`;
-   */
-  const programNode = csf._ast.program;
-  let foundConfigImport = false;
-
-  // Check if a root-level constant named 'config' exists
-  const hasRootLevelConfig = programNode.body.some(
-    (n) =>
-      t.isVariableDeclaration(n) &&
-      n.declarations.some((declaration) => t.isIdentifier(declaration.id, { name: 'config' }))
-  );
-
-  const sbConfigImportName = hasRootLevelConfig ? 'storybookConfig' : 'config';
-
-  const sbConfigImportSpecifier = t.importSpecifier(
-    t.identifier(sbConfigImportName),
-    t.identifier('config')
-  );
-
-  programNode.body.forEach((node) => {
-    if (t.isImportDeclaration(node) && isValidPreviewPath(node.source.value)) {
-      const hasConfigSpecifier = node.specifiers.some(
-        (specifier) =>
-          t.isImportSpecifier(specifier) && t.isIdentifier(specifier.imported, { name: 'config' })
-      );
-
-      if (!hasConfigSpecifier) {
-        node.specifiers.push(sbConfigImportSpecifier);
-      }
-
-      foundConfigImport = true;
-    }
-  });
-
-  const hasMeta = !!csf._meta;
-
-  Object.entries(csf._storyExports).forEach(([key, decl]) => {
-    const id = decl.id;
-    const declarator = decl as t.VariableDeclarator;
-    let init = t.isVariableDeclarator(declarator) ? declarator.init : undefined;
-
-    if (t.isIdentifier(id) && init) {
-      if (t.isTSSatisfiesExpression(init) || t.isTSAsExpression(init)) {
-        init = init.expression;
-      }
-
-      if (t.isObjectExpression(init)) {
-        const typeAnnotation = id.typeAnnotation;
-        // Remove type annotation as it's now inferred
-        if (typeAnnotation) {
-          id.typeAnnotation = null;
-        }
-
-        // Wrap the object in `meta.story()`
-        declarator.init = t.callExpression(
-          t.memberExpression(t.identifier(metaVariableName), t.identifier('story')),
-          [init]
-        );
-      } else if (t.isArrowFunctionExpression(init)) {
-        // Transform CSF1 to meta.story({ render: <originalFn> })
-        const renderProperty = t.objectProperty(t.identifier('render'), init);
-
-        const objectExpression = t.objectExpression([renderProperty]);
-
-        declarator.init = t.callExpression(
-          t.memberExpression(t.identifier(metaVariableName), t.identifier('story')),
-          [objectExpression]
-        );
-      }
-    }
-  });
-
-  // modify meta
-  if (csf._metaPath) {
-    let declaration = csf._metaPath.node.declaration;
-    if (t.isTSSatisfiesExpression(declaration) || t.isTSAsExpression(declaration)) {
-      declaration = declaration.expression;
-    }
-
-    if (t.isObjectExpression(declaration)) {
-      const metaVariable = t.variableDeclaration('const', [
-        t.variableDeclarator(
-          t.identifier(metaVariableName),
-          t.callExpression(
-            t.memberExpression(t.identifier(sbConfigImportName), t.identifier('meta')),
-            [declaration]
-          )
-        ),
-      ]);
-      csf._metaPath.replaceWith(metaVariable);
-    } else if (t.isIdentifier(declaration)) {
-      /**
-       * Transform const declared metas:
-       *
-       * `const meta = {}; export default meta;`
-       *
-       * Into a meta call:
-       *
-       * `const meta = config.meta({ title: 'A' });`
-       */
-      const binding = csf._metaPath.scope.getBinding(declaration.name);
-      if (binding && binding.path.isVariableDeclarator()) {
-        const originalName = declaration.name;
-
-        // Always rename the meta variable to 'meta'
-        binding.path.node.id = t.identifier(metaVariableName);
-
-        let init = binding.path.node.init;
-        if (t.isTSSatisfiesExpression(init) || t.isTSAsExpression(init)) {
-          init = init.expression;
-        }
-        if (t.isObjectExpression(init)) {
-          binding.path.node.init = t.callExpression(
-            t.memberExpression(t.identifier(sbConfigImportName), t.identifier('meta')),
-            [init]
-          );
-        }
-
-        // Update all references to the original name
-        csf._metaPath.scope.rename(originalName, metaVariableName);
-      }
-
-      // Remove the default export, it's not needed anymore
-      csf._metaPath.remove();
-    }
-  }
-
-  if (hasMeta && !foundConfigImport) {
-    const configImport = t.importDeclaration(
-      [sbConfigImportSpecifier],
-      t.stringLiteral('#.storybook/preview')
-    );
-    programNode.body.unshift(configImport);
-  }
-
-  // Remove type imports – now inferred – from @storybook/* packages
-  const disallowlist = [
-    'Story',
-    'StoryFn',
-    'StoryObj',
-    'Meta',
-    'MetaObj',
-    'ComponentStory',
-    'ComponentMeta',
-  ];
-
-  programNode.body = programNode.body.filter((node) => {
-    if (t.isImportDeclaration(node)) {
-      const { source, specifiers } = node;
-
-      if (source.value.startsWith('@storybook/')) {
-        const allowedSpecifiers = specifiers.filter((specifier) => {
-          if (t.isImportSpecifier(specifier) && t.isIdentifier(specifier.imported)) {
-            return !disallowlist.includes(specifier.imported.name);
-          }
-          // Retain non-specifier imports (e.g., namespace imports)
-          return true;
-        });
-
-        // Remove the entire import if no specifiers are left
-        if (allowedSpecifiers.length > 0) {
-          node.specifiers = allowedSpecifiers;
-          return true;
-        }
-
-        // Remove the import if no specifiers remain
-        return false;
-      }
-    }
-
-    // Retain all other nodes
-    return true;
-  });
-
-  return printCsf(csf).code;
-}
-
-const logger = console;
-
-async function runStoriesCodemod(dryRun: boolean | undefined) {
-  try {
-    let globString = 'src/stories/*.stories.*';
-    if (!process.env.IN_STORYBOOK_SANDBOX) {
-      logger.log('Please enter the glob for your stories to migrate');
-      globString = (
-        await prompts({
-          type: 'text',
-          name: 'glob',
-          message: 'glob',
-          initial: globString,
-        })
-      ).glob;
-    }
-    logger.log('Applying codemod on your stories...');
-    await runCodemod(globString, storyToCsfFactory, { dryRun });
-  } catch (err: any) {
-    console.log('err message', err.message);
-    if (err.message === 'No files matched') {
-      console.log('going to run again');
-      await runStoriesCodemod(dryRun);
-    } else {
-      throw err;
-    }
-  }
-}
-
-export const csfFactories: CommandFix = {
-  id: 'csf-factories',
-  promptType: 'command',
-  async run({ dryRun, mainConfig, mainConfigPath, previewConfigPath, packageJson }) {
-    const frameworkPackageName = getFrameworkPackageName(mainConfig);
-
-    const rendererPackageName =
-      (await getRendererPackageNameFromFramework(frameworkPackageName as string)) ??
-      (await detectRenderer(packageJson));
-
-    console.log({ rendererPackageName, frameworkPackageName });
-
-    await runStoriesCodemod(dryRun);
-    logger.log('Applying codemod on your main config...');
-    await runCodemod(mainConfigPath, (fileInfo) =>
-      configToCsfFactory(fileInfo, 'main', { dryRun })
-    );
-    logger.log('Applying codemod on your preview config...');
-    await runCodemod(previewConfigPath, (fileInfo) =>
-      configToCsfFactory(fileInfo, 'preview', { dryRun })
-    );
-  },
-};
