+++ scripts/tasks/sandbox-parts.ts
@@ -819,55 +819,14 @@ export const extendPreview: Task['run'] = async ({ template, sandboxDir }) => {
     template.expected.framework === '@storybook/react-vite' &&
     !template.skipTasks.includes('vitest-integration')
   ) {
-    // add CSF4 style config
-    previewConfig.setImport(['defineConfig'], '@storybook/react/preview');
-    // and all of the addons/previewAnnotations that are needed
     previewConfig.setImport(null, '../src/stories/components');
-    previewConfig.setImport(
-      { namespace: 'addonEssentialsAnnotations' },
-      '@storybook/addon-essentials/entry-preview'
-    );
-    previewConfig.setImport({ namespace: 'addonA11yAnnotations' }, '@storybook/addon-a11y/preview');
-    previewConfig.setImport(
-      { namespace: 'addonActionsAnnotations' },
-      '@storybook/addon-actions/preview'
-    );
-    previewConfig.setImport(
-      { namespace: 'addonTestAnnotations' },
-      '@storybook/experimental-addon-test/preview'
-    );
     previewConfig.setImport({ namespace: 'coreAnnotations' }, '../template-stories/core/preview');
     previewConfig.setImport(
       { namespace: 'toolbarAnnotations' },
       '../template-stories/addons/toolbars/preview'
     );
-
-    previewConfig.setBodyDeclaration(
-      t.exportNamedDeclaration(
-        t.variableDeclaration('const', [
-          t.variableDeclarator(
-            t.identifier('config'),
-            t.callExpression(t.identifier('defineConfig'), [
-              t.objectExpression([
-                t.spreadElement(t.identifier('preview')),
-                t.objectProperty(
-                  t.identifier('addons'),
-                  t.arrayExpression([
-                    t.identifier('addonEssentialsAnnotations'),
-                    t.identifier('addonA11yAnnotations'),
-                    t.identifier('addonActionsAnnotations'),
-                    t.identifier('addonTestAnnotations'),
-                    t.identifier('coreAnnotations'),
-                    t.identifier('toolbarAnnotations'),
-                  ])
-                ),
-              ]),
-            ])
-          ),
-        ]),
-        []
-      )
-    );
+    previewConfig.appendNodeToArray(['addons'], t.identifier('coreAnnotations'));
+    previewConfig.appendNodeToArray(['addons'], t.identifier('toolbarAnnotations'));
   }
 
   if (template.expected.builder.includes('vite')) {
