+++ scripts/tasks/sandbox.ts
@@ -2,6 +2,7 @@ import dirSize from 'fast-folder-size';
 // eslint-disable-next-line depend/ban-dependencies
 import { pathExists, remove } from 'fs-extra';
 import { join } from 'path';
+import prompts from 'prompts';
 import { promisify } from 'util';
 
 import { now, saveBench } from '../bench/utils';
@@ -144,11 +145,11 @@ export const sandbox: Task = {
 
     await extendMain(details, options);
 
-    await extendPreview(details, options);
-
     await setImportMap(details.sandboxDir);
 
     await runMigrations(details, options);
+
+    await extendPreview(details, options);
 
     logger.info(`✅ Storybook sandbox created at ${details.sandboxDir}`);
   },
